python/requirements.txt
aiogram
asyncpg
redis
sqlalchemy
grpcio
grpcio-tools